+++ src/electron/is.js
@@ -1,4 +1,5 @@
 module.exports = {
-  mac: process.platform === 'darwin',
   dev: process.env.NODE_ENV === 'development',
+  mac: process.platform === 'darwin',
+  windows: process.platform === 'win32',
 }