+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add `Ctrl+]` and `Ctrl+]` (macOS: `Cmd+]` and `Cmd+]`) for opening the next and previous track in the "Get Info" window
 - Add metadata editing of bpm, track number, total tracks, disc number and total discs
 - Add more items to the Song menu
+- Add "Reveal in File Manager" menu item
 - Add `Escape` shortcut for deselecting tracks
 - Add support selecting tracks via `Ctrl+A` (macOS: `Cmd+A`), `Shift+Up/Down`, `Alt+Up/Down`, `Shift+Alt+Up/Down`
 - Expand the titlebar drag region a bit